pyproject.toml
[project]
name = "job-description-copilot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "gradio>=6.3.0",
    "langchain>=1.2.6",
    "langchain-groq>=1.1.1",
    "langgraph>=1.0.6",
    "pipeline>=0.1.0",
    "pymupdf>=1.26.7",
    "pypdf>=6.6.0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.53.0",
    "tiktoken>=0.12.0",
]
